pyproject.toml
[tool.poetry]
name = "vet-researcher"
version = "0.1.0"
description = "Agentic pet medical history searching"
authors = ["fourcolors <sterlingcobb@gmail.com>"]
readme = "README.md"
packages = [{ include = "vet_researcher" }]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
swarms = ">=7.2.2,<8.0.0"
swarm-models = ">=0.3.2,<0.4.0"
swarms-memory = ">=0.1.2,<0.2.0"
python-dotenv = "^1.0.1"
click = "^8.1.7"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.5.1"

[tool.poetry.scripts]
vet-researcher = "vet_researcher.cli.main:cli"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
multi_line_output = 3
